Rewrite all tests so we could conduct unit test if we want (For now all tests pass).
