+++ src/test/java/edu/pitt/medschool/service/AnalysisServiceTest.java
@@ -1,10 +1,6 @@
-/**
- * 
- */
 package edu.pitt.medschool.service;
 
-import static org.junit.Assert.assertEquals;
-
+import edu.pitt.medschool.model.dto.Downsample;
 import org.apache.commons.lang3.time.DateFormatUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -12,13 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import edu.pitt.medschool.model.dto.Downsample;
-import edu.pitt.medschool.service.AnalysisService;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 
-/**
- * @author Isolachine
- *
- */
 @RunWith(SpringRunner.class)
 @SpringBootTest
 public class AnalysisServiceTest {
@@ -26,9 +18,13 @@ public class AnalysisServiceTest {
     AnalysisService analysisService;
 
     @Test
-    public void test() {
-        Downsample query = analysisService.selectByPrimaryKey(4);
-        assertEquals("03/27/2018 10:38:12", DateFormatUtils.format(query.getCreateTime(), "MM/dd/yyyy HH:mm:ss"));
+    public void testGet() {
+        Downsample query = analysisService.selectByPrimaryKey(26);
+        assertEquals("07/02/2018 19:21:48", DateFormatUtils.format(query.getCreateTime(), "MM/dd/yyyy HH:mm:ss"));
+        assertEquals("09/17/2018", DateFormatUtils.format(query.getUpdateTime(), "MM/dd/yyyy"));
+        assertEquals((int) query.getPeriod(), 3600);
+        assertEquals((int) query.getDuration(), 172800);
+        assertFalse(query.getDownsampleFirst());
     }
 
 }
