+++ src/test/java/edu/pitt/medschool/service/ImportCSVTest.java
@@ -1,26 +1,27 @@
 package edu.pitt.medschool.service;
 
+import org.junit.Ignore;
+import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
 
-@SpringBootApplication(scanBasePackages = {"edu.pitt.medschool"})
-public class ImportCSVTest implements CommandLineRunner {
+import static org.junit.Assert.assertEquals;
 
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ImportCSVTest {
     @Autowired
     ImportCsvService ics;
+
     @Value("${machine}")
     private String uuid;
 
-    public static void main(String[] args) {
-        SpringApplication.run(ImportCSVTest.class, args);
-    }
-
-    @Override
-    public void run(String... args) throws Exception {
-        ics._test();
+    @Test
+    public void basicTest() throws Exception {
+        assertEquals(ics.GetUUID(), uuid);
     }
 
 }
