--- src/test/java/edu/pitt/medschool/test/TestController.java
@@ -1,12 +1,0 @@
-/**
- * 
- */
-package edu.pitt.medschool.test;
-
-/**
- * @author Isolachine
- *
- */
-public class TestController {
-
-}
